Add basic platform and plugin handlers refactor keyboard
