+++ src/wayland_display.cc
@@ -6,15 +6,8 @@
 #include "wayland_display.h"
 
 #include <linux/input-event-codes.h>
-#include <rapidjson/document.h>
-#include <rapidjson/stringbuffer.h>
-#include <rapidjson/writer.h>
-#include <sys/mman.h>
 #include <sys/types.h>
-#include <sys/wait.h>
 #include <unistd.h>
-#include <xkbcommon/xkbcommon.h>
-
 #include <chrono>
 #include <sstream>
 #include <string>
@@ -112,7 +105,7 @@ void WaylandDisplay::InitializeApplication(
   args.command_line_argv = command_line_args_c.data();
   args.platform_message_callback = [](const FlutterPlatformMessage* message,
                                       void* context) {
-    reinterpret_cast<WaylandDisplay*>(context)->PlatformMessageCallback(
+    reinterpret_cast<WaylandDisplay*>(context)->platform_channel_.PlatformMessageCallback(
         message);
   };
 
@@ -133,22 +126,7 @@ void WaylandDisplay::InitializeApplication(
   custom_task_runners.platform_task_runner = &platform_task_runner;
   args.custom_task_runners = &custom_task_runners;
 
-  static constexpr char kAccessibilityChannel[] = "flutter/accessibility";
-  static constexpr char kFlutterPlatformChannel[] = "flutter/platform";
-  static constexpr char kTextInputChannel[] = "flutter/textinput";
-  static constexpr char kKeyEventChannel[] = "flutter/keyevent";
-  static constexpr char kFlutterPlatformViewsChannel[] = "flutter/platform_views";
-
-  static constexpr char kPluginFlutterIoConnectivity[] = "plugins.flutter.io/connectivity";
-  static constexpr char kPluginFlutterIoUrlLauncher[] = "plugins.flutter.io/url_launcher";
-
-  platform_message_handlers_[kAccessibilityChannel] = std::bind(&WaylandDisplay::OnAccessibilityChannelPlatformMessage, this, std::placeholders::_1);
-  platform_message_handlers_[kFlutterPlatformChannel] = std::bind(&WaylandDisplay::OnFlutterPlatformChannelPlatformMessage, this, std::placeholders::_1);
-  platform_message_handlers_[kTextInputChannel] = std::bind(&WaylandDisplay::OnFlutterTextInputChannelPlatformMessage, this, std::placeholders::_1);
-  platform_message_handlers_[kFlutterPlatformViewsChannel] = std::bind(&WaylandDisplay::OnFlutterPlatformViewsChannelPlatformMessage, this, std::placeholders::_1);
-
-  platform_message_handlers_[kPluginFlutterIoConnectivity] = std::bind(&WaylandDisplay::OnFlutterPluginConnectivity, this, std::placeholders::_1);
-  platform_message_handlers_[kPluginFlutterIoUrlLauncher] = std::bind(&WaylandDisplay::OnFlutterPluginIoUrlLauncher, this, std::placeholders::_1);
+  // register platform channel handlers here
 
   FlutterEngine engine = nullptr;
   auto result = FlutterEngineInitialize(FLUTTER_ENGINE_VERSION, &config, &args,
@@ -157,6 +135,8 @@ void WaylandDisplay::InitializeApplication(
     FLWAY_ERROR << "Could not Initialize the Flutter engine" << std::endl;
     return;
   }
+  platform_channel_.SetEngine(engine_);
+
 
   result = FlutterEngineRunInitialized(engine_);
   if (result != kSuccess) {
@@ -374,105 +354,6 @@ WaylandDisplay::WaylandDisplay(size_t width,
 
   if (has_keyboard) {
     std::cout << "Keyboard Present" << std::endl;
-    keyboard = seat.get_keyboard();
-
-    static keyboard_keymap_format keyb_format =
-        keyboard_keymap_format::no_keymap;
-    static struct xkb_context* xkb_context =
-        xkb_context_new(XKB_CONTEXT_NO_FLAGS);
-    static struct xkb_keymap* keymap = NULL;
-    static struct xkb_state* xkb_state = NULL;
-
-    keyboard.on_keymap() = [&](keyboard_keymap_format format, int fd,
-                               uint32_t size) {
-      keyb_format = format;
-      if (format == keyboard_keymap_format::xkb_v1) {
-        char* keymap_string = reinterpret_cast<char*>(
-            mmap(NULL, size, PROT_READ, MAP_SHARED, fd, 0));
-        xkb_keymap_unref(keymap);
-        keymap = xkb_keymap_new_from_string(xkb_context, keymap_string,
-                                            XKB_KEYMAP_FORMAT_TEXT_V1,
-                                            XKB_KEYMAP_COMPILE_NO_FLAGS);
-        munmap(keymap_string, size);
-        close(fd);
-        xkb_state_unref(xkb_state);
-        xkb_state = xkb_state_new(keymap);
-      }
-    };
-
-    keyboard.on_key() = [&](uint32_t, uint32_t, uint32_t key,
-                            keyboard_key_state state) {
-      if (keyb_format == keyboard_keymap_format::xkb_v1) {
-        xkb_keysym_t keysym = xkb_state_key_get_one_sym(xkb_state, key + 8);
-        uint32_t utf32 = xkb_keysym_to_utf32(keysym);
-        if (utf32) {
-          static constexpr char kChannelName[] = "flutter/keyevent";
-
-          static constexpr char kKeyCodeKey[] = "keyCode";
-          static constexpr char kKeyMapKey[] = "keymap";
-          static constexpr char kLinuxKeyMap[] = "linux";
-          static constexpr char kScanCodeKey[] = "scanCode";
-          static constexpr char kModifiersKey[] = "modifiers";
-          static constexpr char kTypeKey[] = "type";
-          static constexpr char kToolkitKey[] = "toolkit";
-          static constexpr char kGLFWKey[] = "glfw";
-          static constexpr char kUnicodeScalarValues[] = "unicodeScalarValues";
-          static constexpr char kKeyUp[] = "keyup";
-          static constexpr char kKeyDown[] = "keydown";
-
-          rapidjson::Document document;
-          auto& allocator = document.GetAllocator();
-          document.SetObject();
-          document.AddMember(kKeyCodeKey, key, allocator);
-          document.AddMember(kKeyMapKey, kLinuxKeyMap, allocator);
-          document.AddMember(kScanCodeKey, key + 8, allocator);
-          document.AddMember(kModifiersKey, 0, allocator);
-          document.AddMember(kToolkitKey, kGLFWKey, allocator);
-          document.AddMember(kUnicodeScalarValues, utf32, allocator);
-
-          switch (state) {
-            case keyboard_key_state::pressed:
-              document.AddMember(kTypeKey, kKeyDown, allocator);
-              break;
-            case keyboard_key_state::released:
-              document.AddMember(kTypeKey, kKeyUp, allocator);
-              break;
-          }
-
-          rapidjson::StringBuffer buffer;
-          rapidjson::Writer<rapidjson::StringBuffer> writer(buffer);
-          document.Accept(writer);
-          std::cout << buffer.GetString() << std::endl;
-
-          FlutterPlatformMessage message = {};
-          message.struct_size = sizeof(FlutterPlatformMessage);
-          message.channel = kChannelName;
-          message.message =
-              reinterpret_cast<const uint8_t*>(buffer.GetString());
-          message.message_size = buffer.GetSize();
-          message.response_handle = nullptr;
-
-          auto result = FlutterEngineSendPlatformMessage(engine_, &message);
-          if (result != kSuccess)
-            std::cout << "FlutterEngineSendPlatformMessage Result: " << result
-                      << std::endl;
-        } else {
-          char name[64];
-          xkb_keysym_get_name(keysym, name, 64);
-          std::cout << "the key " << name << " was "
-                    << ((state == keyboard_key_state::pressed) ? "pressed"
-                                                               : "released")
-                    << std::endl;
-        }
-      }
-    };
-
-    keyboard.on_modifiers() = [&](uint32_t serial, uint32_t mods_depressed,
-                                  uint32_t mods_latched, uint32_t mods_locked,
-                                  uint32_t group) {
-      xkb_state_update_mask(xkb_state, mods_depressed, mods_latched,
-                            mods_locked, 0, 0, group);
-    };
   }
 
   // intitialize egl
@@ -638,259 +519,5 @@ void WaylandDisplay::PostTaskCallback(FlutterTask task, uint64_t target_time) {
   TaskRunner.push(std::make_pair(target_time, task));
 }
 
-void WaylandDisplay::OnAccessibilityChannelPlatformMessage(const FlutterPlatformMessage* message)
-{
-  std::string msg;
-  msg.assign(reinterpret_cast<const char *>(message->message), message->message_size);
-  FLWAY_LOG << "AccessibilityChannel: " << msg << std::endl;
-}
-
-void WaylandDisplay::OnFlutterPlatformChannelPlatformMessage(const FlutterPlatformMessage* message)
-{
-  rapidjson::Document document;
-  document.Parse(reinterpret_cast<const char*>(message->message), message->message_size);
-  if (document.HasParseError() || !document.IsObject()) {
-    return;
-  }
-
-  auto root = document.GetObject();
-  auto method = root.FindMember("method");
-  if (method == root.MemberEnd() || !method->value.IsString()) {
-    return;
-  }
-
-  std::string msg;
-  msg.assign(reinterpret_cast<const char *>(message->message), message->message_size);
-  FLWAY_LOG << "PlatformChannel: " << method->value.GetString() << std::endl;
-
-#if 0
-  static constexpr char kSetApplicationSwitcherDescription[] = "SystemChrome.setApplicationSwitcherDescription";
-  static constexpr char kSetSystemUiOverlayStyle[] = "SystemChrome.setSystemUIOverlayStyle";
-  static constexpr char kSystemNavigatorPopMethod[] = "SystemNavigator.pop";
-  static constexpr char kGetClipboardDataMethod[] = "Clipboard.getData";
-  static constexpr char kSetClipboardDataMethod[] = "Clipboard.setData";
-  static constexpr char kSystemSoundPlay[] = "SystemSound.play";
-  static constexpr char kTextPlainFormat[] = "text/plain";
-  static constexpr char kTextKey[] = "text";
-#endif
-
-  FlutterEngineSendPlatformMessageResponse(engine_, message->response_handle, nullptr, 0);
-}
-
-void WaylandDisplay::OnFlutterTextInputChannelPlatformMessage(const FlutterPlatformMessage* message)
-{
-  rapidjson::Document document;
-  document.Parse(reinterpret_cast<const char*>(message->message), message->message_size);
-  if (document.HasParseError() || !document.IsObject()) {
-    return;
-  }
-  auto root = document.GetObject();
-  auto method = root.FindMember("method");
-  if (method == root.MemberEnd() || !method->value.IsString()) {
-    return;
-  }
-
-  std::string msg;
-  msg.assign(reinterpret_cast<const char *>(message->message), message->message_size);
-  FLWAY_LOG << "TextInput: " << method->value.GetString() << std::endl;
-
-#if 0
-  if (method->value == "TextInput.show") {
-    if (ime_) {
-      text_sync_service_->ShowKeyboard();
-    }
-  } else if (method->value == "TextInput.hide") {
-    if (ime_) {
-      text_sync_service_->HideKeyboard();
-    }
-  } else 
-
-  if (method->value == "TextInput.setClient") {
-    current_text_input_client_ = 0;
-    DeactivateIme();
-    auto args = root.FindMember("args");
-    if (args == root.MemberEnd() || !args->value.IsArray() ||
-        args->value.Size() != 2)
-      return;
-    const auto& configuration = args->value[1];
-    if (!configuration.IsObject()) {
-      return;
-    }
-    // TODO(abarth): Read the keyboard type from the configuration.
-    current_text_input_client_ = args->value[0].GetInt();
-
-    auto initial_text_input_state = fuchsia::ui::input::TextInputState{};
-    initial_text_input_state.text = "";
-    last_text_state_ = std::make_unique<fuchsia::ui::input::TextInputState>(
-        initial_text_input_state);
-    ActivateIme();
-  } else if (method->value == "TextInput.setEditingState") {
-    if (ime_) {
-      auto args_it = root.FindMember("args");
-      if (args_it == root.MemberEnd() || !args_it->value.IsObject()) {
-        return;
-      }
-      const auto& args = args_it->value;
-      fuchsia::ui::input::TextInputState state;
-      state.text = "";
-      // TODO(abarth): Deserialize state.
-      auto text = args.FindMember("text");
-      if (text != args.MemberEnd() && text->value.IsString())
-        state.text = text->value.GetString();
-      auto selection_base = args.FindMember("selectionBase");
-      if (selection_base != args.MemberEnd() && selection_base->value.IsInt())
-        state.selection.base = selection_base->value.GetInt();
-      auto selection_extent = args.FindMember("selectionExtent");
-      if (selection_extent != args.MemberEnd() &&
-          selection_extent->value.IsInt())
-        state.selection.extent = selection_extent->value.GetInt();
-      auto selection_affinity = args.FindMember("selectionAffinity");
-      if (selection_affinity != args.MemberEnd() &&
-          selection_affinity->value.IsString() &&
-          selection_affinity->value == "TextAffinity.upstream")
-        state.selection.affinity = fuchsia::ui::input::TextAffinity::UPSTREAM;
-      else
-        state.selection.affinity = fuchsia::ui::input::TextAffinity::DOWNSTREAM;
-      // We ignore selectionIsDirectional because that concept doesn't exist on
-      // Fuchsia.
-      auto composing_base = args.FindMember("composingBase");
-      if (composing_base != args.MemberEnd() && composing_base->value.IsInt())
-        state.composing.start = composing_base->value.GetInt();
-      auto composing_extent = args.FindMember("composingExtent");
-      if (composing_extent != args.MemberEnd() &&
-          composing_extent->value.IsInt())
-        state.composing.end = composing_extent->value.GetInt();
-      ime_->SetState(std::move(state));
-    }
-  } else if (method->value == "TextInput.clearClient") {
-    current_text_input_client_ = 0;
-    last_text_state_ = nullptr;
-    DeactivateIme();
-  } else {
-    FLWAY_ERROR << "Unknown " << message->channel << " method "
-                    << method->value.GetString();
-  }
-  #endif
-}
-
-void WaylandDisplay::OnFlutterPlatformViewsChannelPlatformMessage(const FlutterPlatformMessage* message)
-{
-  rapidjson::Document document;
-  document.Parse(reinterpret_cast<const char*>(message->message), message->message_size);
-  if (document.HasParseError() || !document.IsObject()) {
-    FLWAY_ERROR << "Could not parse document";
-    return;
-  }
-  auto root = document.GetObject();
-  auto method = root.FindMember("method");
-  if (method == root.MemberEnd() || !method->value.IsString()) {
-    return;
-  }
-
-  std::string msg;
-  msg.assign(reinterpret_cast<const char *>(message->message), message->message_size);
-  FLWAY_LOG << "PlatformViews: " << method->value.GetString() << std::endl;
-
-  if (method->value == "View.enableWireframe") {
-    auto args_it = root.FindMember("args");
-    if (args_it == root.MemberEnd() || !args_it->value.IsObject()) {
-      FLWAY_ERROR << "No arguments found.";
-      return;
-    }
-    const auto& args = args_it->value;
-
-    auto enable = args.FindMember("enable");
-    if (!enable->value.IsBool()) {
-      FLWAY_ERROR << "Argument 'enable' is not a bool";
-      return;
-    }
-
-    FLWAY_LOG << "wireframe_enabled_callback goes here" << std::endl;
-  } else {
-    FLWAY_ERROR << "Unknown " << message->channel << " method "
-                    << method->value.GetString();
-  }
-}
-
-void WaylandDisplay::OnFlutterPluginIoUrlLauncher(const FlutterPlatformMessage* message)
-{
-  std::unique_ptr<std::vector<std::uint8_t>> result;
-  auto codec = &flutter::StandardMethodCodec::GetInstance();
-  auto method_call = codec->DecodeMethodCall(message->message, message->message_size);
-
-  if (method_call->method_name().compare("launch") == 0) {
-    std::string url;
-    if (method_call->arguments() && method_call->arguments()->IsMap()) {
-      const EncodableMap &arguments = method_call->arguments()->MapValue();
-      auto url_it = arguments.find(EncodableValue("url"));
-      if (url_it != arguments.end()) {
-        url = url_it->second.StringValue();
-      }
-    }
-    if (url.empty()) {
-      auto result = codec->EncodeErrorEnvelope("argument_error", "No URL provided");
-      goto done;
-    }
-
-    pid_t pid = fork();
-    if (pid == 0) {
-      execl("/usr/bin/xdg-open", "xdg-open", url.c_str(), nullptr);
-      exit(1);
-    }
-    int status = 0;
-    waitpid(pid, &status, 0);
-    if (status != 0) {
-      std::ostringstream error_message;
-      error_message << "Failed to open " << url << ": error " << status;
-      result = codec->EncodeErrorEnvelope("open_error", error_message.str());
-      goto done;
-    }
-    auto val = EncodableValue(true);
-    result = codec->EncodeSuccessEnvelope(&val);
-
-  } else if (method_call->method_name().compare("canLaunch") == 0) {
-    std::string url;
-    if (method_call->arguments() && method_call->arguments()->IsMap()) {
-      const EncodableMap &arguments = method_call->arguments()->MapValue();
-      auto url_it = arguments.find(EncodableValue("url"));
-      if (url_it != arguments.end()) {
-        url = url_it->second.StringValue();
-      }
-    }
-    if (url.empty()) {
-      result = codec->EncodeErrorEnvelope("argument_error", "No URL provided");
-      goto done;
-    }
-    flutter::EncodableValue response(
-      (url.rfind("https:", 0) == 0) || (url.rfind("http:", 0) == 0) ||
-      (url.rfind("ftp:", 0) == 0) || (url.rfind("file:", 0) == 0));
-    result = codec->EncodeSuccessEnvelope(&response);
-  }
-done:
-  FlutterEngineSendPlatformMessageResponse(engine_, message->response_handle, result->data(), result->size());
-}
-
-void WaylandDisplay::OnFlutterPluginConnectivity(const FlutterPlatformMessage* message)
-{
-  std::string msg;
-  msg.assign(reinterpret_cast<const char *>(message->message), message->message_size);
-  FLWAY_LOG << "Connectivity: " << msg << std::endl;
-
-}
-
-void WaylandDisplay::PlatformMessageCallback(
-    const FlutterPlatformMessage* message) {
-
-  // Find the handler for the channel; if there isn't one, report the failure.
-  if (platform_message_handlers_.find(message->channel) == platform_message_handlers_.end()) {
-    FlutterEngineSendPlatformMessageResponse(engine_, message->response_handle,
-                                        nullptr, 0);
-    return;
-  }
-
-  auto& message_handler = platform_message_handlers_[message->channel];
-
-  message_handler(message);
-}
 
 }  // namespace flutter
